+++ GOVKit/Sources/GOVKit/GroupedList/GroupedListHeader.swift
@@ -1,0 +1,13 @@
+import Foundation
+import UIKit
+
+public struct GroupedListHeader {
+    public let title: String
+    let icon: UIImage?
+
+    public init(title: String,
+                icon: UIImage?) {
+        self.title = title
+        self.icon = icon
+    }
+}
