+++ GOVKit/Sources/GOVKit/GroupedList/GroupedListRow.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public protocol GroupedListRow {
+    var id: String { get }
+    var title: String { get }
+    var body: String? { get }
+}
+
+public extension GroupedListRow {
+    var body: String? {
+        nil
+    }
+}
