+++ GOVKit/Sources/GOVKit/GroupedList/GroupedListSection.swift
@@ -15,136 +15,6 @@ public struct GroupedListSection {
     }
 }
 
-public struct GroupedListHeader {
-    public let title: String
-    let icon: UIImage?
-    
-    public init(title: String,
-                icon: UIImage?) {
-        self.title = title
-        self.icon = icon
-    }
-}
-
-public protocol GroupedListRow {
-    var id: String { get }
-    var title: String { get }
-    var body: String? { get }
-}
-
-public extension GroupedListRow {
-    var body: String? {
-        nil
-    }
-}
-
-public struct LinkRow: GroupedListRow,
-                       Identifiable {
-    public let id: String
-    public let title: String
-    public let body: String?
-    public var isWebLink: Bool = true
-    public let action: () -> Void
-    
-    public init(id: String,
-                title: String,
-                body: String? = nil,
-                isWebLink: Bool = true,
-                action: @escaping () -> Void) {
-        self.id = id
-        self.title = title
-        self.body = body
-        self.isWebLink = isWebLink
-        self.action = action
-    }
-}
-
-public class DetailRow: GroupedListRow,
-                        Identifiable {
-    public let id: String
-    public let title: String
-    public let body: String
-    public let destructive: Bool
-    public let accessibilityHint: String
-    public let action: () -> Void
-
-    public init(id: String,
-                title: String,
-                body: String,
-                accessibilityHint: String,
-                destructive: Bool = false,
-                action: @escaping () -> Void) {
-        self.id = id
-        self.title = title
-        self.body = body
-        self.accessibilityHint = accessibilityHint
-        self.destructive = destructive
-        self.action = action
-    }
-}
-
-
-public struct NavigationRow: GroupedListRow,
-                             Identifiable {
-    public let id: String
-    public let title: String
-    public let body: String?
-    let action: () -> Void
-    
-    public init(id: String,
-                title: String,
-                body: String?,
-                action: @escaping () -> Void) {
-        self.id = id
-        self.title = title
-        self.body = body
-        self.action = action
-    }
-}
-
-public struct InformationRow: GroupedListRow,
-                              Identifiable {
-    public let id: String
-    public let title: String
-    public let body: String?
-    public let imageName: String?
-    let detail: String
-    
-    public init(id: String,
-                title: String,
-                body: String?,
-                imageName: String? = nil,
-                detail: String) {
-        self.id = id
-        self.title = title
-        self.body = body
-        self.imageName = imageName
-        self.detail = detail
-    }
-}
-
-public class ToggleRow: GroupedListRow,
-                        ObservableObject {
-    public var id: String
-    public let title: String
-    @Published var isOn: Bool {
-        didSet {
-            self.action(isOn)
-        }
-    }
-    let action: ((Bool) -> Void)
-
-    public init(id: String,
-         title: String,
-         isOn: Bool,
-         action: @escaping (Bool) -> Void) {
-        self.title = title
-        self.isOn = isOn
-        self.action = action
-        self.id = id
-    }
-}
-
 public struct GroupedListSection_Previews: PreviewProvider {
     public static var previews: some View {
         Text("preview")
